pyproject.toml
[project]
name = "neon"
version = "0.1.0"
description = "Gaze analysis with interactive velocity plot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "scipy>=1.10",
    "pupil-labs-realtime-api>=1.1.0",
]

